renamed userAction to profileActions
